[tests/tfmma.c] Fixed type in printf.
[From SVN r9799 (trunk)]

+++ tests/tfmma.c
@@ -42,7 +42,7 @@ random_test (mpfr_t a, mpfr_t b, mpfr_t c, mpfr_t d, mpfr_rnd_t rnd)
   inex_ref = mpfr_add (ref, ab, cd, rnd);
   if (inex_res != inex_ref)
     {
-      printf ("mpfr_fmma failed for p=%u rnd=%d\n", p, rnd);
+      printf ("mpfr_fmma failed for p=%ld rnd=%d\n", (long int) p, rnd);
       printf ("a="); mpfr_dump (a);
       printf ("b="); mpfr_dump (b);
       printf ("c="); mpfr_dump (c);
@@ -62,7 +62,7 @@ random_test (mpfr_t a, mpfr_t b, mpfr_t c, mpfr_t d, mpfr_rnd_t rnd)
   inex_ref = mpfr_sub (ref, ab, cd, rnd);
   if (inex_res != inex_ref)
     {
-      printf ("mpfr_fmms failed for p=%u rnd=%d\n", p, rnd);
+      printf ("mpfr_fmms failed for p=%ld rnd=%d\n", (long int) p, rnd);
       printf ("a="); mpfr_dump (a);
       printf ("b="); mpfr_dump (b);
       printf ("c="); mpfr_dump (c);
